refactor(OAuthUserInfo): 디렉토리 이동 #21
- 도메인에서 쓰이는 객체이기에 vo 패키지로 이동
